% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex.R
\name{check_sex}
\alias{check_sex}
\alias{predict_sex}
\title{Infer the sex of sample donor}
\usage{
check_sex(raw)

predict_sex(X, Y, male, female)
}
\arguments{
\item{raw}{Output of calling \code{\link{read_idats()}}}

\item{X, Y}{Forwarded from \code{check_sex()}}

\item{male, female}{Indices of male and female samples}
}
\value{
\code{check_sex} returns the normalized average total intensities of probes targeting the X and Y chromosomes. These are forwarded to \code{predict_sex} which returns a factor with levels "f","m" (and \code{NA} if the sex cannot be determined conclusively).
}
\description{
Return the normalized average total intensities of probes targeting the X and Y chromosomes.
}
