% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_metrics.R
\name{control_metrics}
\alias{control_metrics}
\alias{sample_failure}
\title{Quality control metrics}
\usage{
control_metrics(raw)

sample_failure(metrics)
}
\arguments{
\item{raw}{Output of calling \code{\link{read_idats}}}

\item{metrics}{Output of callling \code{control_metrics}}
}
\value{
For \code{control_metrics}, a list of 17 control metrics, each one a numeric vector equal in length to the sample size

For \code{sample_failure}, a logical vector, \code{TRUE} if sample at corresponding index failed on any of the 17 control metrics.
}
\description{
Calculate the QC metrics as described in the 'BeadArray Controls Reporter Software Guide' from Illumina.
}
\note{
You can download the Software Guide at https://support.illumina.com/downloads/beadarray-controls-reporter-software-guide-1000000004009.html
}
