% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLC.R
\name{estimateLC}
\alias{estimateLC}
\title{Estimate leukcoyte composition}
\usage{
estimateLC(meth)
}
\value{
Estimated cell proportions B-lymphocytes, CD4 T-cells, CD8 T-cells, Granulocytes, Monocytes and Natural killer cells using the Houseman algorithm. The model was trained on a dataset of both peripheral and cord blood, in order to apply to study populations of broad age range.
}
\description{
Estimate leukcoyte composition
}
\references{
{Houseman EA, Accomando WP, Koestler DC, Christensen BC, Marsit CJ, Nelson HH, Wiencke JK, Kelsey KT. DNA methylation arrays as surrogate measures of cell mixture distribution. BMC bioinformatics. 2012 Dec 1;13(1):86.}
}
