% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_scores.R
\name{methylation_score}
\alias{methylation_score}
\title{Various methylation scores}
\usage{
methylation_score(beta, model)
}
\arguments{
\item{beta}{matrix of beta values}

\item{model}{Name of the model, i.e., one of `maternal_smoking`, `gestational_age`,
or `horvath_clock`}
}
\description{
Various methylation scores
}
