% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionP.R
\name{summits}
\alias{summits}
\title{Return peaks of beta-value distribution}
\usage{
summits(beta)
}
\description{
Return the locations of the two peaks (completly methylated and unmethylated CpG sites)
of the beta-value distribution.
}
