% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes.R
\name{call_genotypes}
\alias{call_genotypes}
\alias{mxm_}
\alias{snp_outliers}
\alias{eBeta}
\title{Genotype calling}
\usage{
call_genotypes(snpmatrix, learn = FALSE, maxiter = 50)

mxm_(genotypes)

snp_outliers(genotypes)

eBeta(x, w)
}
\arguments{
\item{snpmatrix}{Matrix of beta-values for SNP probes. Provide SNPs probes as rows and samples as columns.}

\item{maxiter}{Maximal number of iterations of the Expectation-Maximization algorithm learning the mixture model}

\item{genotypes}{Output of \code{call_genotypes}}
}
\value{
For \code{call_genotypes}, a list containing
\item{par}{Parameters of the mixture model}
\item{loglik}{Log-likelihood in each iteration of the EM algorithm}
\item{outliers}{A-posteriori probability of SNP being an outlier}
\item{gamma}{A-posteriori probabilities for each of the three genotypes}

For \code{snp_outliers}, a metric assessing the outlierness of the SNP beta-values. High values may indicate either contaminated or failed samples.

For \code{mxm_}, a histogram showing the distribution of beta-values for SNP probes with the density function of the mixture model overlaid.
}
\description{
Detect SNP probes which do not fit into on of the three categories (AA,AB,BB).
A mixture model (3 Beta distributions, 1 uniform distribution for outliers) is fitted to all SNP probes.
After learning the model parameters via EM algorithm, the probability of being an outlier is computed for each SNP.
}
\author{
Jonathan A. Heiss
}
