% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_agreement.R
\name{check_snp_agreement}
\alias{check_snp_agreement}
\alias{agreement_}
\alias{enumerate_sample_donors}
\title{Agreement of SNPs}
\usage{
check_snp_agreement(genotypes, donor_ids, sample_ids)

agreement_(genotypes, donor_ids, sample_ids, ...)

enumerate_sample_donors(genotypes)
}
\arguments{
\item{genotypes}{Output of \code{call_genotypes}}

\item{donor_ids}{Vector of donor IDs, should include duplicates}

\item{sample_ids}{Vector of unique sample IDs}
}
\value{
\code{check_snp_agreement} returns all conflicts found between user-supplied donor IDs
and those derived from the genetic fingerprint. Either \code{NULL} if not conflicts were found,
or a list of \code{data.table}s, each representing one connected component of the conflict graph.

\code{enumerate_sample_donors} returns for each sample the donor IDs.
}
\description{
These function check the agreement of genetic fingerprints between samples
(after genotype calling with \code{call_genotypes}). In case of \code{check_snp_agreement},
the user supplies for each sample the assumed donor ID and the function returns all conflicts,
meaning either samples that are supposed to come from the same donor but have distinct genetic
fingerprints, or samples from supposedly different donors with the same fingerprint.
\code{enumerate_sample_donors} on the other hand is inferring donor IDs in case they are unknown
(useful for example to detect technical replicates in a public dataset).
}
\author{
Jonathan Heiss
}
