% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_agreement.R
\name{check_snp_agreement}
\alias{check_snp_agreement}
\alias{agreement_}
\title{Agreement of SNPs}
\usage{
check_snp_agreement(genotypes, weights, donor_ids, sample_ids)

agreement_(genotypes, weights, donor_ids, sample_ids)
}
\arguments{
\item{genotypes}{A list of length 3, each element a matrix with probablilities}

\item{weights}{A matrix of weights (1-probability of being an outlier) for each SNP}

\item{donor_ids}{Vector of donor IDs, should include duplicates}

\item{sample_ids}{Vector of unique sample IDs}
}
\value{
\code{check_snp_agreement} returns all connnected components of the conflict graph as list of data.tables. \code{agreement_} computes the same scores but returns a plot instead.
}
\description{
Compute relative agreement scores of SNP probes. Beta-values from the probes interrogating SNPs (59 for EPIC, 65 for 450K) are binned into three categories. Subsequently, conflicts are listed, either samples from supposedly same donor with a low agreement between SNP probes, or samples from supposedly different donors with a suspiciously high agreement.
}
\author{
Jonathan A. Heiss
}
