% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionP.R
\name{detectionP}
\alias{detectionP}
\alias{detectionP.neg}
\alias{mask}
\alias{eval_detP_cutoffs}
\alias{detectionP.minfi}
\title{Detection p-values}
\usage{
detectionP(raw)

detectionP.neg(raw)

mask(raw, threshold)

eval_detP_cutoffs(raw, males = NULL, females = NULL)

detectionP.minfi(rgSet)
}
\arguments{
\item{raw}{Output of calling \code{\link{read_idats()}}, must include component \code{detP} for
\code{mask()} and \code{eval_detP_threshold()}.}

\item{threshold}{p-value threshold (arithmetic scale) above which oberservations are set to NA.}

\item{rgSet}{minfi rgSet object}

\item{male/female}{Indices of male and female subjects}
}
\value{
For \code{detectionP()} and \code{detectionP.neg()} a modified \code{raw} object with a
\code{detP} component, a matrix of detection p-values, added. \code{detectionP()} computes p-values
on the linear scale, whereas \code{detectionP.neg()} returns p-values on the log10 scale.

For \code{detectionP.minfi()} a matrix of detection p-values.

For \code{mask()}, a modified \code{raw} object, with undetected probes set to \code{NA}.
}
\description{
Detection p-values are based on the distribution of the intensities
of the negative control probes or the U (M) intensities observed for completely methylated
(unmethylated) probes, respectively.

\code{detectionP()} is the recommend approach that generates realistic p-values as described in Heiss and Just, 2019.

\code{detectionP.neg()} follows the approach used in GenomeStudio (p-values are unrealistic).

\code{detectionP.minfi()} provides an implementation for \code{RGChannelSet} objects as used in the \code{minfi} package.

\code{eval_detP_threshold()} generates a plot showing the number of undetected chrY probes among
male and female subjects for various p-value thresholds, in order to empirically choose a threshold.

Finally, \code{mask()} is masking all probes with detection p-values below the specified threshold.
}
\references{
{Heiss JA, Just AC. Improved filtering of DNA methylation microarray data by detection
p-values and its impact on downstream analyses. Clinical Epigenetics (2019) 11:15}
}
\author{
Jonathan A. Heiss
}
