% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{drop_samples}
\alias{drop_samples}
\title{Drop samples from raw data.}
\usage{
drop_samples(raw, j = NULL)
}
\arguments{
\item{raw}{Output of calling \code{\link{read_idats}}}

\item{j}{Indices of the samples to drop}
}
\value{
A modified \code{raw} object
}
\description{
Drop samples from raw data.
}
\author{
Jonathan A. Heiss
}
