% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLC.R
\name{estimateLC}
\alias{estimateLC}
\title{Estimate leukcoyte composition}
\usage{
estimateLC(meth, ref)
}
\value{
Estimated cell proportions B-lymphocytes, CD4 T-cells, CD8 T-cells, Granulocytes, Monocytes and Natural killer cells using the Houseman algorithm. The model was trained on a dataset of both peripheral and cord blood, in order to apply to study populations of broad age range.
}
\description{
Estimate leukcoyte composition
}
\references{
{Houseman EA, Accomando WP, Koestler DC, Christensen BC, Marsit CJ, Nelson HH, Wiencke JK, Kelsey KT. DNA methylation arrays as surrogate measures of cell mixture distribution. BMC bioinformatics. 2012 Dec 1;13(1):86.}

{Gervin K, Page CM, Aass HC, Jansen MA, Fjeldstad HE, Andreassen BK, Duijts L, van Meurs JB, van Zelm MC, Jaddoe VW, Nordeng H. Cell type specific DNA methylation in cord blood: A 450K-reference data set and cell count-based validation of estimated cell type composition. Epigenetics. 2016 Sep 1;11(9):690-8}

{Reinius LE, Acevedo N, Joerink M, Pershagen G, Dahlén SE, Greco D, Söderhäll C, Scheynius A, Kere J. Differential DNA methylation in purified human blood cells: implications for cell lineage and studies on disease susceptibility. PloS one. 2012 Jul 25;7(7):e41361.}

{de Goede OM, Razzaghian HR, Price EM, Jones MJ, Kobor MS, Robinson WP, Lavoie PM. Nucleated red blood cells impact DNA methylation and expression analyses of cord blood hematopoietic cells. Clinical epigenetics. 2015 Dec;7(1):95.}
}
