% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{Preprocessing}
\alias{Preprocessing}
\alias{correct_dye_bias}
\alias{correct_dye_bias2}
\alias{normalize}
\alias{dont_normalize}
\title{Functions for preprocessing}
\usage{
correct_dye_bias(raw)

correct_dye_bias2(raw)

normalize(raw, tissue = "", quiet = FALSE)

dont_normalize(raw)
}
\arguments{
\item{raw}{Output of calling \code{\link{read_idats}}}

\item{tissue}{Optional. If set to "blood", predefined reference values are used for normalization. Recommened when cell proportions are to be estimated.}
}
\value{
A modified \code{raw} object with dye-bias corrected intensities using RELIC for \code{correct_dye_bias}. A matrix of beta-values, either normalized (for \code{normalize}) or not (for \code{dont_normalize}).
}
\description{
Functions for preprocessing
}
\references{
{Xu Z, Langie SA, De Boever P, Taylor JA, Niu L. RELIC: a novel dye-bias correction method for Illumina Methylation BeadChip. BMC Genomics. 2017 Jan 3;18(1):4.}

{Heiss JA, Brenner H. Between-array normalization for 450K data. Frontiers in Genetics. 2015;6.}
}
\author{
Jonathan Heiss
}
