% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_idats}
\alias{read_idats}
\title{Read .idat files}
\usage{
read_idats(idat_files, quiet = FALSE, on_disk = FALSE)
}
\arguments{
\item{idat_files}{Character vector of relative or absolute filepaths, but without the suffixes
_Grn.idat' and '_Red.idat'. IDATs for red and green channel must have the same prefix and be
stored in the same folder. E.g., a sample with the idats 200607110235_R01C01_Red.idat and
200607110235_R01C01_Grn.idat would be passed to \code{read_idats} as "200607110235_R01C01".}

\item{quiet}{If \code{TRUE}, suppresses the progress bar (useful for RMarkdown scripts).}

\item{on_disk}{If \code{TRUE}, data will be stored on disk using the ff package.
This will be slower but enables processing datasets larger than memory.}
}
\value{
A list containing
\item{manifest}{A data.table describing the probes}
\item{M}{intensities of targeting methylated sequences}
\item{U}{intensities of targeting unmethylated sequences}
\item{N}{number of beads from which average intensities in M were derived}
\item{V}{number of beads from which average intensities in U were derived}
\item{ctrlG}{Intensities of the control probes in the green color channel}
\item{ctrlR}{Intensities of the control probes in the red color channel}
\item{meta}{A data.table containing unique sample_ids and metadata}
}
\description{
Read .idat files
}
\examples{
\dontrun{
read_idats('9976861004_R01C01')
}

}
