% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_idats}
\alias{read_idats}
\title{Read .idat files}
\usage{
read_idats(idat_files, quiet = FALSE)
}
\arguments{
\item{idat_files}{Character vector of relative or absolute filepaths, but without the suffixes '_Grn.idat' and '_Red.idat'}

\item{quiet}{If TRUE, a progress bar is shown.}
}
\value{
A list containing
\item{manifest}{A data.table describing the probes}
\item{M}{intensities of targeting methylated sequences}
\item{U}{intensities of targeting unmethylated sequences}
\item{N}{number of beads from which average intensities in M were derived}
\item{V}{number of beads from which average intensities in U were derived}
\item{ctrlG}{Intensities of the control probes in the green color channel}
\item{ctrlR}{Intensities of the control probes in the red color channel}
\item{meta}{A data.table containing unique sample_ids and metadata}
}
\description{
Read .idat files
}
\examples{
\dontrun{
read_idats('9976861004_R01C01')
}

}
